src/requirements.txt
# V5 AI Engine Requirements
# Complete standalone dependencies

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Async Support
aiohttp==3.9.1
asyncio==3.4.3
asyncpg==0.29.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography==41.0.7

# Database
asyncpg==0.29.0
redis[hiredis]==5.0.1
sqlalchemy==2.0.23

# AI/ML
llama-cpp-python==0.2.20
numpy==1.24.3
scikit-learn==1.3.2

# Input Validation
jsonschema==4.20.0
email-validator==2.1.0
python-dotenv==1.0.0

# Rate Limiting & Caching
redis[hiredis]==5.0.1
cachetools==5.3.2

# Monitoring & Observability
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-exporter-otlp==1.21.0
prometheus-client==0.19.0

# Utilities
pyyaml==6.0.1
python-json-logger==2.0.7
tenacity==8.2.3
httpx==0.25.2

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.0
mypy==1.7.1
coverage==7.3.3

# Documentation
mkdocs==1.5.3
mkdocs-material==9.5.2